ImageValFinetune/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0
peft>=0.4.0
trl>=0.4.7
bitsandbytes>=0.41.0
scipy
pandas
openpyxl
pillow>=9.0.0
tqdm
numpy
scikit-learn
einops
flamingo-pytorch  # Add this for Flamingo implementation